+++ nakadion/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.28.13] - 2020-10-23
+
+### CHANGED
+
+- The committer will commit in a fixed interval if "after seconds" is set with the `CommitStragtegy`
+
 ## [0.28.12] - 2020-09-30
 
 ### CHANGED